Fix more warnings

+++ src/camera.h
@@ -55,7 +55,7 @@ typedef CameraEvent CutsceneShot;
 
 void func_8001CA10(Camera*);
 void func_8001CA24(Player*, f32);
-void func_8001CA78(Player*, Camera*, Vec3f, f32*, f32*, f32*, s16, s8);
+void func_8001CA78(Player*, Camera*, Vec3f, f32*, f32*, f32*, s32, s32);
 void func_8001D53C(Player*, Camera*, Vec3f, f32*, f32*, f32*, s16, s16);
 void func_8001D794(Player*, Camera*, Vec3f, f32*, f32*, f32*, s16);
 void func_8001E0C4(Camera*, Player*, s8);
